Restructure files and rename folders
