+++ src/components/PatientProfile/TriagePills.tsx
