+++ src/components/ScanPatient/BarcodeScan.tsx
